[backend] refactor: move recommendations preview class DynamicWebsitePreviewRecommendations into his own file
